# [Perf improvement] GTiff: avoid using block cache when writing whole blocks