--- a/port/cpl_vsi_mem.cpp
+++ b/port/cpl_vsi_mem.cpp
@@ -48,6 +48,20 @@
 #include <utility>
 #include <memory>
 
+#include <mutex>
+// c++17 or VS2017
+#if __cplusplus >= 201703L || _MSC_VER >= 1910
+#include <shared_mutex>
+#define CPL_SHARED_MUTEX_TYPE std::shared_mutex
+#define CPL_SHARED_LOCK std::shared_lock<std::shared_mutex>
+#define CPL_EXCLUSIVE_LOCK std::unique_lock<std::shared_mutex>
+#else
+// Poor-man implementation of std::shared_mutex with an exclusive mutex
+#define CPL_SHARED_MUTEX_TYPE std::mutex
+#define CPL_SHARED_LOCK std::lock_guard<std::mutex>
+#define CPL_EXCLUSIVE_LOCK std::lock_guard<std::mutex>
+#endif
+
 #include "cpl_atomic_ops.h"
 #include "cpl_conv.h"
 #include "cpl_error.h"
@@ -65,10 +79,7 @@
 ** want to create and read different files at the same time and so might
 ** collide access oFileList without the mutex.
 **
-** VSIMemFile: In theory we could allow different threads to update the
-** the same memory file, but for simplicity we restrict to single writer,
-** multiple reader as an expectation on the application code (not enforced
-** here), which means we don't need to do any protection of this class.
+** VSIMemFile: A mutex protects accesses to the file
 **
 ** VSIMemHandle: This is essentially a "current location" representing
 ** on accessor to a file, and is inherently intended only to be used in
@@ -77,10 +88,8 @@
 ** In General:
 **
 ** Multiple threads accessing the memory filesystem are ok as long as
-**  1) A given VSIMemHandle (i.e. FILE * at app level) isn't used by multiple
-**     threads at once.
-**  2) A given memory file isn't accessed by more than one thread unless
-**     all threads are just reading.
+** a given VSIMemHandle (i.e. FILE * at app level) isn't used by multiple
+** threads at once.
 */
 
 /************************************************************************/
@@ -105,6 +114,7 @@ class VSIMemFile
     vsi_l_offset nMaxLength = GUINTBIG_MAX;
 
     time_t mTime = 0;
+    CPL_SHARED_MUTEX_TYPE m_oMutex{};
 
     VSIMemFile();
     virtual ~VSIMemFile();
@@ -223,6 +233,7 @@ VSIMemFile::~VSIMemFile()
 /*                             SetLength()                              */
 /************************************************************************/
 
+// Must be called under exclusive lock
 bool VSIMemFile::SetLength(vsi_l_offset nNewLength)
 
 {
@@ -327,6 +338,8 @@ int VSIMemHandle::Close()
 int VSIMemHandle::Seek(vsi_l_offset nOffset, int nWhence)
 
 {
+    CPL_SHARED_LOCK oLock(poFile->m_oMutex);
+
     bExtendFileAtNextWrite = false;
     if (nWhence == SEEK_CUR)
     {
@@ -380,6 +393,8 @@ vsi_l_offset VSIMemHandle::Tell()
 size_t VSIMemHandle::Read(void *pBuffer, size_t nSize, size_t nCount)
 
 {
+    CPL_SHARED_LOCK oLock(poFile->m_oMutex);
+
     size_t nBytesToRead = nSize * nCount;
     if (nBytesToRead == 0)
         return 0;
@@ -417,6 +432,8 @@ size_t VSIMemHandle::Read(void *pBuffer, size_t nSize, size_t nCount)
 size_t VSIMemHandle::PRead(void *pBuffer, size_t nSize,
                            vsi_l_offset nOffset) const
 {
+    CPL_SHARED_LOCK oLock(poFile->m_oMutex);
+
     if (nOffset < poFile->nLength)
     {
         const size_t nToCopy = static_cast<size_t>(
@@ -436,6 +453,8 @@ size_t VSIMemHandle::PRead(void *pBuffer, size_t nSize,
 size_t VSIMemHandle::Write(const void *pBuffer, size_t nSize, size_t nCount)
 
 {
+    CPL_EXCLUSIVE_LOCK oLock(poFile->m_oMutex);
+
     if (!bUpdate)
     {
         errno = EACCES;
@@ -496,6 +515,8 @@ int VSIMemHandle::Truncate(vsi_l_offset nNewSize)
     }
 
     bExtendFileAtNextWrite = false;
+
+    CPL_EXCLUSIVE_LOCK oLock(poFile->m_oMutex);
     if (poFile->SetLength(nNewSize))
         return 0;
 
@@ -579,6 +600,7 @@ VSIVirtualHandle *VSIMemFilesystemHandler::Open(const char *pszFilename,
     // Overwrite
     else if (strstr(pszAccess, "w"))
     {
+        CPL_EXCLUSIVE_LOCK oLock(poFile->m_oMutex);
         poFile->SetLength(0);
         poFile->nMaxLength = nMaxLength;
     }
@@ -605,7 +627,10 @@ VSIVirtualHandle *VSIMemFilesystemHandler::Open(const char *pszFilename,
              pszFilename, static_cast<int>(poFile.use_count()));
 #endif
     if (strstr(pszAccess, "a"))
+    {
+        CPL_SHARED_LOCK oLock(poFile->m_oMutex);
         poHandle->m_nOffset = poFile->nLength;
+    }
 
     return poHandle;
 }
@@ -631,16 +656,18 @@ int VSIMemFilesystemHandler::Stat(const char *pszFilename,
         return 0;
     }
 
-    if (oFileList.find(osFilename) == oFileList.end())
+    auto oIter = oFileList.find(osFilename);
+    if (oIter == oFileList.end())
     {
         errno = ENOENT;
         return -1;
     }
 
-    std::shared_ptr<VSIMemFile> poFile = oFileList[osFilename];
+    std::shared_ptr<VSIMemFile> poFile = oIter->second;
 
     memset(pStatBuf, 0, sizeof(VSIStatBufL));
 
+    CPL_SHARED_LOCK oLock(poFile->m_oMutex);
     if (poFile->bIsDirectory)
     {
         pStatBuf->st_size = 0;
@@ -676,18 +703,19 @@ int VSIMemFilesystemHandler::Unlink_unlocked(const char *pszFilename)
 {
     const CPLString osFilename = NormalizePath(pszFilename);
 
-    if (oFileList.find(osFilename) == oFileList.end())
+    auto oIter = oFileList.find(osFilename);
+    if (oIter == oFileList.end())
     {
         errno = ENOENT;
         return -1;
     }
 
 #ifdef DEBUG_VERBOSE
-    std::shared_ptr<VSIMemFile> poFile = oFileList[osFilename];
+    std::shared_ptr<VSIMemFile> poFile = oIter->second;
     CPLDebug("VSIMEM", "Unlink %s: ref_count=%d (before)", pszFilename,
              static_cast<int>(poFile.use_count()));
 #endif
-    oFileList.erase(oFileList.find(osFilename));
+    oFileList.erase(oIter);
 
     return 0;
 }