… autotest/gcore/tiff_write.py