This avoids less contention on the block cache mutex and saves memory
copies, particular when writing dataset-level RasterIO() with a buffer
with data for all bands, and with a pixel interleaving.

The optimization triggers if the following conditions are met:
- (xsize, ysize) = (bufxsize, bufysize), ie no resampling
- origin (xoff, yoff) of buffer to write is aligned on a tile boundary
- end (xoff+xsize, yoff+ysize) of buffer is aligned on tile boundary or
  raster right/bottom
- only for bit depth = 8, 16, 32, 64 or 128
- the block cache for the bands to considered has not been initialized
  at all
- and for dataset level write with pixel interleaving, all bands are
  written at once.

The visible side effects of this optimization are hopefully few, but one
particular case exhibited by the test_tiff_ovr_10() case is when writing
with a lossy compression and re-reading the raster without closing it in
between. Previously if the raster was small enough to fit in the block
cache, re-reading it would get the original pixel values. With this
change, we now get the values after decompression.

Using the gtiff_multi_ds_parallel_write.py program to write 2048x2048
tiled datasets in parallel from 12 threads, ZSTD compression

$ time python ../perftests/gtiff_multi_ds_parallel_write.py --nbands 1 --without-optim
real    0m17,768s
user    2m55,418s
sys     0m1,174s

$ time python ../perftests/gtiff_multi_ds_parallel_write.py --nbands 1
real    0m9,149s
user    1m34,882s
sys     0m0,597s

$ time python ../perftests/gtiff_multi_ds_parallel_write.py --nbands 3 --buffer-interleave BAND --without-optim
real    0m16,106s
user    2m42,623s
sys     0m0,834s

$ time python ../perftests/gtiff_multi_ds_parallel_write.py --nbands 3 --buffer-interleave BAND
real    0m9,013s
user    1m31,537s
sys     0m0,329s

$ time python ../perftests/gtiff_multi_ds_parallel_write.py --nbands 3 --buffer-interleave PIXEL --without-optim
real    0m41,574s
user    3m23,985s
sys     3m45,286s

$ time python ../perftests/gtiff_multi_ds_parallel_write.py --nbands 3 --buffer-interleave PIXEL
real    0m5,893s
user    0m56,796s
sys     0m0,402s

Helps for #4272
CC @tbonfort