@tbonfort 2023-05-12
wow! the pixel interleaved multiband case speedup is impressive, and the two other ones are also very significant. can't wait to test this out. :clap: @rouault