@tbonfort 2023-05-12
> The visible side effects of this optimization are hopefully few, but one
> particular case exhibited by the test_tiff_ovr_10() case is when writing
> with a lossy compression and re-reading the raster without closing it in
> between. Previously if the raster was small enough to fit in the block
> cache, re-reading it would get the original pixel values. With this
> change, we now get the values after decompression.

This is good to know but not unexpected: it will happen all the time with this PR applied whereas it could happen or not beforehand depending on the fullness of the blockcache. i.e. this was not some behavior that could be relied on.