@tbonfort 2023-05-13
@rouault I was wondering, will this new codepath get activated on a regular gdal_translate and/or gdalwarp command?