@rouault 2023-05-13
> will this new codepath get activated on a regular gdal_translate and/or gdalwarp command?

gdal_translate, definitely yes, as shown by the change of behavior of test_tiff_ovr_10() which uses CreateCopy(), and that's what gdal_translate ultimately does.

For gdalwarp, not by default unless you specify the OPTIMIZE_SIZE=YES warping option (``-wo OPTIMIZE_SIZE=YES``). Cf https://gdal.org/api/gdalwarp_cpp.html#_CPPv4N15GDALWarpOptions16papszWarpOptionsE : "OPTIMIZE_SIZE: This defaults to FALSE, but may be set to TRUE typically when writing to a compressed dataset (GeoTIFF with COMPRESSED creation option set for example) for achieving a smaller file size. This is achieved by writing at once data aligned on full blocks of the target dataset, which avoids partial writes of compressed blocks and lost space when they are rewritten at the end of the file. However sticking to target block size may cause major processing slowdown for some particular reprojections."

This reminds me that I wanted to try improving that default behaviour of gdalwarp to enable OPTIMIZE_SIZE when it is reasonable to do so (now filed as https://github.com/OSGeo/gdal/issues/7761)