@tbonfort 2023-05-13
so am I right in thinking that with this patch there might be some marginal gains when using gdal_translate, and noticeable ones when using gdalwarp -multi -wo optimize_size ?