@rouault 2023-05-13
> there might be some marginal gains when using gdal_translate, and noticeable ones when using gdalwarp -multi -wo optimize_size ?

Hard to anticipate gains as those work flows are not pure I/O. gdalwarp should be generally more CPU intensive in non I/O code paths that gdalwarp, so I'm not sure the effects of this PR will be very visible. 
As you mention the -multi flag of gdalwarp, I guess you might be confused by what it does. It has generally very modest influence as it does ReadRaster() and WriteRaster() in a single thread, separated from the one(s) that execute the computational part of the warping (coordinate reprojection + resampling kernel). You won't get N threads writing to the same dataset, which isn't something by the GDAL data model anyway.