@tbonfort 2023-05-15
LGTM: I have seen a ~50% decrease in processing time for workflows with multiple readers and a single writer